+++ source/planet/terrain/terrain_generation.h
@@ -27,7 +27,7 @@ int _corner_type (const Terrain&, const Corner*);
 int _edge_type (const Terrain&, const Edge*);
 void _classify_terrain (Terrain&, const Grid&);
 
-float _elevation_at_point (const Vector3&, const std::vector<std::array<Vector3, 3> >&);
+double _elevation_at_point (const Vector3&, const std::vector<std::array<Vector3, 3> >&);
 
 //returns point on sphere of uniform distribution, given two random integers
 Vector3 point_uniform (int, int);
